# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alimentalos API",
        "description": "Official Alimentalos API Documentation",
        "contact": {
            "email": "iantorres@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v.0.6-beta"
    },
    "servers": [
        {
            "url": "https://api.alimentalos.cl",
            "description": "Backend production server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users.",
                "description": "Returns the users paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user.",
                "description": "Returns the recently created user instance as JSON Object.",
                "operationId": "createUser",
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get groups.",
                "description": "Returns the groups paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getGroups",
                "responses": {
                    "200": {
                        "description": "Groups retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Create group.",
                "description": "Returns the recently created group instance as JSON Object.",
                "operationId": "createGroup",
                "responses": {
                    "200": {
                        "description": "Group created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get alerts.",
                "description": "Returns the alerts paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getAlerts",
                "responses": {
                    "200": {
                        "description": "Alerts retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Create alert.",
                "description": "Returns the recently created alert instance as JSON Object.",
                "operationId": "createAlert",
                "responses": {
                    "200": {
                        "description": "Alert created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get devices.",
                "description": "Returns the devices paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getDevicesPaginated",
                "responses": {
                    "200": {
                        "description": "Devices retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Create device.",
                "description": "Returns the recently created device instance as JSON Object.",
                "operationId": "createDevice",
                "responses": {
                    "200": {
                        "description": "Device created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/geofences": {
            "get": {
                "tags": [
                    "Geofences"
                ],
                "summary": "Get geofences.",
                "description": "Returns the geofences paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getGeofences",
                "responses": {
                    "200": {
                        "description": "Geofences retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Geofences"
                ],
                "summary": "Create geofence.",
                "description": "Returns the recently created geofence instance as JSON Object.",
                "operationId": "createGeofence",
                "responses": {
                    "200": {
                        "description": "Geofence created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/pets": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Get pets.",
                "description": "Returns the pets paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getPets",
                "responses": {
                    "200": {
                        "description": "Pets retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pets"
                ],
                "summary": "Create pet.",
                "description": "Returns the recently created pet instance as JSON Object.",
                "operationId": "createPet",
                "responses": {
                    "200": {
                        "description": "Pet created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/actions": {
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Get actions.",
                "description": "Returns the actions paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getActions",
                "responses": {
                    "200": {
                        "description": "Actions retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/photos": {
            "get": {
                "tags": [
                    "Photos"
                ],
                "summary": "Get photos.",
                "description": "Returns the photos paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getPhotos",
                "responses": {
                    "200": {
                        "description": "Photos retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/pet/locations": {
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create location of pet.",
                "description": "Returns the recently created location of pet as JSON Object.",
                "operationId": "createPetLocationInstance",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Pet personal access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pet location created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/user/locations": {
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create location of user.",
                "description": "Returns the recently created location of user as JSON Object.",
                "operationId": "createUserLocationInstance",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User personal access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User location instance created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/device/locations": {
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create location of device.",
                "description": "Returns the recently created location of device as JSON Object.",
                "operationId": "createDeviceLocationInstance",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User personal access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device location instance created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/groups/{group}/pets": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get pets of group.",
                "description": "Returns the pets of group paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getGroupPets",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Unique identifier of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pets retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/users/{user}/pets": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get pets of user.",
                "description": "Returns the pets of user paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getUserPets",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Unique identifier of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pets retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}